refactor: move price chart to components/common fix: skeleton width
