+++ apps/dapp/src/pages/ssov-beta/[[...slug]].tsx
@@ -6,9 +6,9 @@ import { NextSeo } from 'next-seo';
 import useVaultStore from 'hooks/ssov/useVaultStore';
 
 import PageLayout from 'components/common/PageLayout';
+import PriceChart from 'components/common/PriceChart';
 import AsidePanel from 'components/ssov-beta/AsidePanel';
 import InfoBox from 'components/ssov-beta/InfoBox';
-import PriceChart from 'components/ssov-beta/PriceChart';
 import Positions from 'components/ssov-beta/Tables/Positions';
 import StrikesChain from 'components/ssov-beta/Tables/StrikesChain';
 import TitleBar from 'components/ssov-beta/TitleBar';
@@ -85,9 +85,10 @@ const SsovBetaMarket = () => {
         />
         <div className="flex space-x-0 lg:space-x-6 flex-col sm:flex-col md:flex-col lg:flex-row space-y-3 md:space-y-0 justify-center">
           <div className="flex flex-col space-y-3 sm:w-full lg:w-3/4 h-full">
-            <div className="h-[520px] rounded-lg text-center flex flex-col justify-center text-stieglitz">
-              <PriceChart market={selectedMarket} />
-            </div>
+            <PriceChart
+              className="h-[520px] rounded-lg text-center flex flex-col justify-center text-stieglitz"
+              market={selectedMarket}
+            />
             <div className="space-y-4">
               <StrikesChain market={selectedMarket} />
               <Positions />
